x Removing warnings
git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@4297 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfo/File__Analyze_MinimizeSize.h
@@ -1211,7 +1211,7 @@ public :
         void    Ibi_Stream_Finish           (int64u Numerator, int64u Denominator); //Partial
         void    Ibi_Add                     ();
     #else //MEDIAINFO_IBI
-        size_t  Ibi_Read_Buffer_Seek        (size_t, int64u, int64u)            {return -1;}
+        size_t  Ibi_Read_Buffer_Seek        (size_t, int64u, int64u)            {return (size_t)-1;}
         void    Ibi_Read_Buffer_Unsynched   ()                                  {}
         void    Ibi_Stream_Finish           ()                                  {}
         void    Ibi_Stream_Finish           (int64u, int64u)                    {}
